+++ Sources/Networking/Networking+Logger.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// A name space for networking
-enum Networking {
+public enum Networking {
     /// A logger associated with data tasks
     internal static let logger: Logger = Logging.frameworkLoggerFactory(category: "Networking")
 }